# Remove some unnecessary string copies in hosting layer