--- a/src/native/corehost/hostmisc/pal.h
+++ b/src/native/corehost/hostmisc/pal.h
@@ -252,7 +252,7 @@ namespace pal
     string_t get_timestamp();
 
     bool getcwd(string_t* recv);
-    string_t to_lower(const string_t& in);
+    string_t to_lower(const char_t* in);
 
 
     inline void file_flush(FILE *f) { std::fflush(f); }