Cut out a bunch of unnecessary string copies:
- Avoid string building for logging versions when logging isn't enabled
- Switch to `pal::char_t*` for some static string literals 
- Switch to `pal::char_t*` for `pal::to_lower` (that's actually what the callers had - we were just forcing them all to create an extra string for no reason)

On Windows, this was about ~3k fewer allocations for launching a console app.

Tried some really rough printf timing on Windows - not sure the best way to do this - I just used `std::chrono::system_clock` to add print lines and averaged over 30 launches. So all grains of salt with these numbers, but it seemed a bit improved on my machine:
| | `corehost_main` to `coreclr_t::create` before loading coreclr (us) | `hostfxr_main_startupinfo` to `coreclr_t::create` before loading coreclr (us) |
| --- | ---: | ---: |
| before | 15,710.33 | 18,556.77 |
| after | 13,611.60 | 16,600.07 |